splintered lib.rs
